refactor(loading): extract adapt logic (#227)
* refactor(loading): extract adapt logic

* chore: remove set-style util totally
